Changes location of transpose method
