+++ tests/test_matrix.py
@@ -1,6 +1,7 @@
 from visma.matrix.checks import isMatrix, dimCheck, multiplyCheck, isSquareMatrix
-from visma.matrix.operations import simplifyMatrix, addMatrix, scalarAdd, scalarSub, scalarMult, scalarDiv, transposeMat
+from visma.matrix.operations import simplifyMatrix, addMatrix, scalarAdd, scalarSub, scalarMult, scalarDiv
 from tests.tester import getTokens
+from visma.matrix.structure import Matrix
 
 ####################
 # matrix.structure #
@@ -165,20 +166,6 @@ def test_scalarDivMatrix():
     assert matSum.__str__() == "[{8.0},{6.0};{4.0},{-0.5}]"
 
 
-def test_transposeMat():
-
-    mat = getTokens("[1, 2; \
-                    3,  4]")
-    matTranspose = transposeMat(mat)
-    assert matTranspose.__str__() == "[{1.0},{3.0};{2.0},{4.0}]"
-
-    mat = getTokens("[5,8,2;\
-                      12,30,9;\
-                      4,17,7]")
-    matTranspose = transposeMat(mat)
-    assert matTranspose.__str__() == "[{5.0},{12.0},{4.0};{8.0},{30.0},{17.0};{2.0},{9.0},{7.0}]"
-
-
 def test_multiplyMatrix():
     """
     # FIXME: Fixing addition fixes multiplication
@@ -199,3 +186,22 @@ def test_multiplyMatrix():
     # assert matPro.__str__() == ""
     """
     pass
+
+
+#################
+# matrix.structure #
+#################
+
+
+def test_transposeMat():
+
+    mat = getTokens("[1, 2; \
+                    3,  4]")
+    matTranspose = Matrix.transposeMat(mat)
+    assert matTranspose.__str__() == "[{1.0},{3.0};{2.0},{4.0}]"
+
+    mat = getTokens("[5,8,2;\
+                      12,30,9;\
+                      4,17,7]")
+    matTranspose = Matrix.transposeMat(mat)
+    assert matTranspose.__str__() == "[{5.0},{12.0},{4.0};{8.0},{30.0},{17.0};{2.0},{9.0},{7.0}]"
