+++ visma/matrix/operations.py
@@ -220,21 +220,3 @@ def scalarDiv(const, mat):
         return matRes
     else:
         logger.error("ZeroDivisionError: Cannot divide matrix by zero")
-
-
-def transposeMat(mat):
-    """Returns Transpose of Matrix
-
-    Arguments:
-        mat {visma.matrix.structure.Matrix} -- matrix token
-
-    Returns:
-        matRes {visma.matrix.structure.Matrix} -- result matrix token
-    """
-    matRes = Matrix()
-    matRes.empty([mat.dim[0], mat.dim[1]])
-    for i in range(mat.dim[0]):
-        for j in range(mat.dim[1]):
-            matRes.value[j][i] = mat.value[i][j]
-    matRes = simplifyMatrix(matRes)
-    return matRes
