+++ visma/matrix/structure.py
@@ -83,6 +83,22 @@ class Matrix(object):
         self.dim[0] = len(self.value)
         self.dim[1] = len(self.value[0])
 
+    def transposeMat(mat):
+        """Returns Transpose of Matrix
+
+        Arguments:
+            mat {visma.matrix.structure.Matrix} -- matrix token
+
+        Returns:
+            matRes {visma.matrix.structure.Matrix} -- result matrix token
+        """
+        matRes = Matrix()
+        matRes.empty([mat.dim[0], mat.dim[1]])
+        for i in range(mat.dim[0]):
+            for j in range(mat.dim[1]):
+                matRes.value[j][i] = mat.value[i][j]
+        return matRes
+
 
 class SquareMat(Matrix):
